moves parser to separate file with comment
