+++ tokens/config.js
@@ -1,39 +1,13 @@
 module.exports = {
   source: ["tokens/*.json"],
-  parsers: [{
-    pattern: /\.json$/,
-    parse: ({ filePath, contents, ...rest }) => {
-      contents = JSON.parse(contents);
-      const categories = Object.entries(contents);
-      for (const [category, categoryValue] of categories) {
-        const types = Object.entries(categoryValue);
-
-        for (const [type, typeValue] of types) {
-          const items = Object.entries(typeValue);
-
-          for (const [item, itemValue] of items) {
-            if (["gradient", "elevation"].includes(type) && itemValue && !itemValue.type) {
-              const subitems = Object.values(itemValue);
-              contents[category][type][item] = {
-                ...subitems[0],
-                extensions: itemValue.extensions
-              };
-
-              contents[category][type][item].value = subitems.filter(v => v && v.value).map(v => v.value);
-            }
-          }
-        }
-      }
-      return contents;
-    }
-  }],
+  parsers: [require("./transformation/common/leoParser")],
   platforms: {
     tailwind: {
       transformGroup: "tailwind/css",
       buildPath: "build/tailwind/",
       files: [
         {
-          destination: "_variables.css",
+          destination: "variables.css",
           format: "tailwind/css",
           filter: "filterWeb",
           options: {
